requirements.txt
requests
Flask==0.12
rq
matplotlib
tensorflow
pillow